_learning_materials/examples/single_agent_basic/requirements.txt
langgraph
langchain
langchain-google-genai
langchain-community
tavily-python
python-dotenv
